cleanup(ROADMAP §6): Group 4 LOW — type & numeric hygiene (per NAMING-STYLE-STANDARD)
Resolve the documented G4 (type & numeric) LOW findings, each behavior-neutral on
the real-AADR golden gate. No §3.2 protected parity vocabulary renamed; no POD/member
underscore changes; no public-API rename. Skipped pointer-wrapper + const __restrict__
buckets and the deferred device-side kernels untouched.

WHAT / WHY (one finding per file):
- cmd_qpgraph.cpp: cast the std::tolower ARGUMENT through unsigned char (the existing
  cast was on the result; signed-char input to <cctype> is UB).
- cmd_qpwave.cpp: debug-only STEPPE_ASSERT making the R0/right-non-empty invariant
  locally self-evident at the right_n subtraction (validated 50+ lines earlier).
- config_builder.cpp: fail-fast range gates for the two qpgraph knobs (--numstart >= 1,
  --max-nadmix >= 0), matching the file's documented input-gate contract.
- qpfstats_jackknife.hpp: paired-length precondition assert (arr/bl both length n_block)
  in the CPU-oracle f2blocks_pair_est.
- dates.cpp: round in long + bound-check each grid dim FITS int before the narrowing cast
  (pathological maxdis/binsize/qbin could otherwise overflow the int these dims index).
- qpfstats.cpp: fail-fast (throw, RELEASE-effective) BEFORE the npopcomb size_t->int
  narrow so the design-matrix sizing cannot consume a wrapped int.
- device_partial_impl.cuh: document the size_t-widening requirement at the resident-buffer
  [P*P*n_block_local] extent (the product lives in the consumer; doc-only here).
- f2_blocks_out.cu: debug-only block-index bounds assert at the tile reader entry.
- qpgraph_fit_kernels.cuh: widen make_layout slab-offset accumulators + dbl_total/int_total
  to long (parity-sensitive: struct passed by value into a parity-gated kernel).
- geno_reader.cpp: apply the file's standard checked-multiply guard to n_lines.
- ind_reader.cpp: stop counting past-cap rows so n_individuals_total honors its
  "== TGENO records" contract (field has no production consumer; gate-confirmed inert).

GATE (the box, Release -Werror, full recompile of all 11 files): default ctest 74/74 green
(GPU-vs-golden, real AADR golden_fit0/fit1_NRBIG/rot); STEPPE_THOROUGH qpadm 7/7 + qpgraph
5/5 green (CpuBackend oracle bit-identical). No regression, no synthetic data.

+++ src/device/cuda/qpgraph_fit_kernels.cuh
@@ -45,17 +45,17 @@ struct QpGraphDeviceTopo {
 struct ScratchLayout {
     int pwts_c, ppwts, Wm, cc, ccs, sc, q1, bl, res, path_w, qf;  // double sub-array offsets
     int nn_w, nn_Ap, nn_qp, nn_z, nn_lu, nn_y;                    // NNLS / solve scratch
-    int dbl_total;                                                // total doubles per thread
+    long dbl_total;                                               // total doubles per thread
     int nn_P, nn_piv, nn_pass;                                    // int sub-array offsets
-    int int_total;                                                // total ints per thread
+    long int_total;                                               // total ints per thread
 };
 
 /// Build the per-thread scratch layout for a topology of (npop,nedge,npair,npath). __host__
 /// __device__: the host sizes the slab + the kernel reconstructs per-topology offsets.
 __host__ __device__ inline ScratchLayout make_layout(int npop, int nedge, int npair, int npath) {
     ScratchLayout L{};
-    int o = 0;
-    auto take = [&](int sz) { int s = o; o += sz; return s; };
+    long o = 0;
+    auto take = [&](int sz) { long s = o; o += sz; return s; };
     L.pwts_c = take(nedge * (npop - 1));
     L.ppwts  = take(npair * nedge);
     L.Wm     = take(npair * nedge);
@@ -74,8 +74,8 @@ __host__ __device__ inline ScratchLayout make_layout(int npop, int nedge, int np
     L.nn_lu  = take(nedge * nedge);
     L.nn_y   = take(nedge);
     L.dbl_total = o;
-    int oi = 0;
-    auto takei = [&](int sz) { int s = oi; oi += sz; return s; };
+    long oi = 0;
+    auto takei = [&](int sz) { long s = oi; oi += sz; return s; };
     L.nn_P    = takei(nedge);
     L.nn_piv  = takei(nedge);
     L.nn_pass = takei(nedge);
